Implement RFC-7 JavaScript module loader in CMS
