+++ .gitignore
@@ -1,2 +1,3 @@
 .sass-cache
 .DS_Store
+node_modules
