+++ code/controllers/CMSMain.php
@@ -78,22 +78,19 @@ class CMSMain extends LeftAndMain implements CurrentPageIdentifier, PermissionPr
 
 		Requirements::css(CMS_DIR . '/css/screen.css');
 		Requirements::customCSS($this->generatePageIconsCss());
-
-		Requirements::combine_files(
-			'cmsmain.js',
-			array_merge(
-				array(
-					CMS_DIR . '/javascript/CMSMain.js',
-					CMS_DIR . '/javascript/CMSMain.EditForm.js',
-					CMS_DIR . '/javascript/CMSMain.AddForm.js',
-					CMS_DIR . '/javascript/CMSPageHistoryController.js',
-					CMS_DIR . '/javascript/CMSMain.Tree.js',
-					CMS_DIR . '/javascript/SilverStripeNavigator.js',
-					CMS_DIR . '/javascript/SiteTreeURLSegmentField.js'
-				),
-				Requirements::add_i18n_javascript(CMS_DIR . '/javascript/lang', true, true)
-			)
-		);
+		Requirements::add_i18n_javascript(CMS_DIR . '/javascript/lang', true, true);
+		Requirements::javascript(CMS_DIR . '/javascript/dist/bundle-lib.js', [
+			'provides' => [
+				CMS_DIR . '/javascript/dist/CMSMain.AddForm.js',
+				CMS_DIR . '/javascript/dist/CMSMain.EditForm.js',
+				CMS_DIR . '/javascript/dist/CMSMain.js',
+				CMS_DIR . '/javascript/dist/CMSMain.Tree.js',
+				CMS_DIR . '/javascript/dist/CMSPageHistoryController.js',
+				CMS_DIR . '/javascript/dist/RedirectorPage.js',
+				CMS_DIR . '/javascript/dist/SilverStripeNavigator.js',
+				CMS_DIR . '/javascript/dist/SiteTreeURLSegmentField.js'
+			]
+		]);
 
 		CMSBatchActionHandler::register('publish', 'CMSBatchAction_Publish');
 		CMSBatchActionHandler::register('unpublish', 'CMSBatchAction_Unpublish');
