+++ javascript/src/bundles/lib.js
@@ -1,0 +1,8 @@
+require('../../src/CMSMain.AddForm.js');
+require('../../src/CMSMain.EditForm.js');
+require('../../src/CMSMain.js');
+require('../../src/CMSMain.Tree.js');
+require('../../src/CMSPageHistoryController.js');
+require('../../src/RedirectorPage.js');
+require('../../src/SilverStripeNavigator.js');
+require('../../src/SiteTreeURLSegmentField.js');
