+++ package.json
@@ -1,0 +1,47 @@
+{
+  "name": "silverstripe-cms",
+  "version": "4.0.0",
+  "description": "The SilverStripe CMS",
+  "directories": {
+    "test": "tests"
+  },
+  "engines": {
+    "node": "~4.2.0"
+  },
+  "scripts": {
+    "build": "gulp build"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/silverstripe/silverstripe-cms.git"
+  },
+  "keywords": [
+    "silverstripe",
+    "cms"
+  ],
+  "author": "SilverStripe Ltd",
+  "license": "BSD-3-Clause",
+  "bugs": {
+    "url": "https://github.com/silverstripe/silverstripe-cms/issues"
+  },
+  "homepage": "https://github.com/silverstripe/silverstripe-cms#readme",
+  "devDependencies": {
+    "babel-core": "^6.4.0",
+    "babel-plugin-transform-es2015-modules-umd": "^6.4.0",
+    "babel-preset-es2015": "^6.3.13",
+    "babelify": "^7.2.0",
+    "browserify": "^13.0.0",
+    "event-stream": "^3.3.2",
+    "glob": "^6.0.4",
+    "gulp": "^3.9.0",
+    "gulp-babel": "^6.1.1",
+    "gulp-diff": "^1.0.0",
+    "gulp-notify": "^2.2.0",
+    "gulp-uglify": "^1.5.1",
+    "gulp-util": "^3.0.7",
+    "semver": "^5.1.0",
+    "vinyl-buffer": "^1.0.0",
+    "vinyl-source-stream": "^1.1.0",
+    "watchify": "^3.7.0"
+  }
+}
